+++ src/cli_shared/cli/mod.rs
@@ -58,9 +58,9 @@ pub struct CliOpts {
     /// Address used for RPC. By defaults binds on localhost on port 2345.
     #[arg(long)]
     pub rpc_address: Option<SocketAddr>,
-    /// P2P addresses, e.g., `--p2p-address /ip4/0.0.0.0/tcp/12345 --p2p-address /ip4/0.0.0.0/tcp/12346`
+    /// P2P listening addresses, e.g., `--p2p-listening-address /ip4/0.0.0.0/tcp/12345 --p2p-listening-address /ip4/0.0.0.0/tcp/12346`
     #[arg(long)]
-    pub p2p_address: Option<Vec<Multiaddr>>,
+    pub p2p_listening_address: Option<Vec<Multiaddr>>,
     /// Allow Kademlia (default: true)
     #[arg(short, long)]
     pub kademlia: Option<bool>,
@@ -178,7 +178,7 @@ impl CliOpts {
             }
         }
 
-        if let Some(addresses) = &self.p2p_address {
+        if let Some(addresses) = &self.p2p_listening_address {
             cfg.network.listening_multiaddrs = addresses.clone();
         }