+++ CHANGELOG.md
@@ -37,8 +37,9 @@
   `forest-tool shed peer-id-from-key-pair`.
 - [#3981](https://github.com/ChainSafe/forest/issues/3981) Add
   `forest-tool backup create|restore`.
-- [#3995](https://github.com/ChainSafe/forest/pull/3995) Add `--p2p-address`
-  option to `forest` to override p2p addresses that forest listens on
+- [#3995](https://github.com/ChainSafe/forest/pull/3995) Add
+  `--p2p-listening-address` option to `forest` to override p2p addresses that
+  forest listens on
 
 ### Changed